Do not size allocations from an unverified region extent
An adversarial review found a remote memory DoS introduced by this branch.

Inside an open region an offset is validated against the allowance rather than
against bytes that exist, and pushing a limit for a non-trailing dynamic field
turns the open region into a bounded one. LengthKnown reported that bounded
extent as known, so the field's own decode sized an allocation from it before
reading any payload: 12 crafted bytes made the decoder allocate ~500 MB, held
for as long as the peer kept the connection open. ssz-max does not help, since
the real spec limit for ExecutionPayload transactions is above the allowance.

LengthKnown now means the extent is backed by input known to exist, which is
only true once the stream length is established -- at EOF, or because the caller
supplied it. A limit still bounds reads; it just cannot be trusted as an
allocation size, so those regions are consumed incrementally. Peak heap for the
reproducer drops from 477 MiB to 0.3 MiB.

DecodeRemaining's incremental loop now also stops at the region limit rather
than only at EOF: a bounded-but-unverified region would otherwise refill for
bytes belonging to the next region and never make progress.

Also from the review:
  - GrowSlice returned nil for an empty list where ExpandSlice deliberately
    returns an empty slice, so an empty trailing list decoded to nil on the
    codegen unknown-size path and to [] everywhere else.
  - WithMaxStreamSize(math.MaxInt) overflowed the read-one-past-the-allowance
    arithmetic and rejected every payload; values above 2^32 truncated in the
    uint32 conversions callers apply to a region length. The allowance is now
    capped, which an SSZ offset cannot exceed anyway.
  - Generated code read the trailing-element condition through itemCount, which
    a nested list shadows between the push and the pop, so the two compared
    against different lists.

The region helpers move into unmarshal.go alongside ExpandSlice and
DecodeUint64Slice: the decoder_ prefix is for decoder implementations, not for
the helpers generated code calls.
