+++ CHANGELOG.md
@@ -15,6 +15,10 @@ All notable changes to the `dynamic-ssz` library are documented here.
   limits are now enforced incrementally while reading rather than after buffering.
 - `Decoder` gains `LengthKnown`, `More`, `DecodeRemaining` and `PushOpenLimit`;
   `sszutils.NewUnknownStreamDecoder` constructs a decoder for a payload of unknown length.
+  `LengthKnown` reports whether a region's extent is backed by input known to exist, not
+  merely whether a limit is set: inside an open region an offset is validated against the
+  allowance, so a limit derived from one can declare far more bytes than were delivered and
+  must never be used as an allocation size.
 
 ### Changed
 - Errors from an unknown-size decode are less precise: checks that a known length catches
@@ -25,6 +29,8 @@ All notable changes to the `dynamic-ssz` library are documented here.
 - The `DynamicViewUnmarshaler` reflection path now uses the type's fixed size when it has
   one, matching the fastssz and `DynamicUnmarshaler` paths, instead of always consuming the
   whole remaining region.
+- Generated decoders no longer read the trailing-element condition through a name a nested
+  list can shadow, which made the push and the pop compare against two different lists.
 
 ### Upgrade notes
 - Types whose SSZ methods were generated by v1.3.2 or earlier must be **regenerated** to be
